tests/wpt/tests/navigation-api/state/same-document-away-and-back-navigation-api.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(async () => {
    let start_length = navigation.entries().length;
    let entry0 = navigation.currentEntry;
    let start_index = entry0.index;

    let navState = { statevar: "state" };
    await navigation.navigate("#1", { state: navState }).committed;
    let entry1 = navigation.currentEntry;

    navState.startvar2 = "otherstate";
    await navigation.navigate("#2", { state: navState }).committed;
    let entry2 = navigation.currentEntry;

    t.step_func(() => {
      assert_equals(navigation.entries().length, start_length + 2);
      assert_equals(entry0, navigation.entries()[start_index]);
      assert_equals(entry1, navigation.entries()[start_index + 1]);
      assert_equals(entry2, navigation.entries()[start_index + 2]);

      assert_equals(entry0.getState(), undefined);

      let state1 = entry1.getState();
      assert_not_equals(state1, undefined);
      assert_not_equals(state1, navState);
      assert_equals(state1.statevar, "state");
      assert_equals(state1.startvar2, undefined);

      let state2 = entry2.getState();
      assert_not_equals(state2, undefined);
      assert_not_equals(state2, navState);
      assert_equals(state2.statevar, "state");
      assert_equals(state2.startvar2, "otherstate");

      navigation.back();
      window.onpopstate = t.step_func_done(() => {
        assert_equals(navigation.entries().length, start_length + 2);
        let back_entry = navigation.currentEntry;
        assert_equals(back_entry, navigation.entries()[start_index + 1]);
        let back_state = back_entry.getState();
        assert_not_equals(back_state, state1);
        assert_not_equals(back_state, state2);
        assert_equals(back_state.statevar, "state");
        assert_equals(back_state.startvar2, undefined);
      });
    })();
  }, 0);
}, "entry.getState() behavior after navigating away using the navigation API, then back");
</script>
